+++ UserControl.API/Requests/GetUsersRequest.cs
@@ -1,3 +1,3 @@
 namespace UserControl.API.Requests;
 
-public record GetUsersRequest(int[] UsersIds);
+public record GetUsersRequest(long[] UsersIds);
